Rename check.proto to checkconfig.proto
